feat: modularized function to upload any file to S3
