+++ package.json
@@ -31,7 +31,8 @@
     "morgan": "^1.10.0",
     "node-mime-types": "^1.1.0",
     "pg": "^8.8.0",
-    "sequelize": "^6.21.4"
+    "sequelize": "^6.21.4",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
     "eslint": "^8.22.0",
